Added OBJ loading

+++ src/material.cuh
@@ -183,7 +183,7 @@ struct dielectric : public material {
     __device__ inline static float reflectance(float cosine, float ref_idx) {
         float r0 = (1.0f - ref_idx) / (1.0f + ref_idx);
         r0 = r0 * r0;
-        return r0 + (1.0f - r0) * pow(1.0f - cosine, 5.0f);
+        return r0 + (1.0f - r0) * powf(1.0f - cosine, 5.0f);
     }
 };
